Artix install update
Former-commit-id: c64a4384a0e641bc13f9caf556061352cc2dcb0a [formerly 78cd867c7e5ee14aefce7b4f831dc005fecfae4a]
Former-commit-id: 8ec101fd7124be3ae5233dbbf4bd2d569397b6a1
Former-commit-id: 7e2d55954eb5df1d3b5f6bceafc73f6ea1ab174f

+++ artix/after-chroot.sh
@@ -79,7 +79,6 @@ confirm "Install base packages?"
 rm -rf /usr/lib64
 doas -u $USER1 sh $ARTIXD_DIR/install-base.sh
 
-
 confirm "Install packages?"
 PACKAGE_LIST=''
 for group in "${PACKAGE_GROUPS[@]}"; do
@@ -120,13 +119,15 @@ rc-update del agetty.tty1 default
 rc-update add agetty-autologin.tty1 default
 
 
-pri "Installing dotfiles for user $USER1"
-rm -rf $USER_HOME/.config
-export USER1
-doas -u $USER1 sh $DOTFILES_DIR/install-dotfiles.sh
+if [ $INSTALL_DOTFILES -eq 1 ]; then
+    pri "Installing dotfiles for user $USER1"
+    rm -rf $USER_HOME/.config
+    export USER1
+    doas -u $USER1 sh $DOTFILES_DIR/install-dotfiles.sh
 
-pri "Installing dotfiles for root"
-sh $DOTFILES_DIR/install-dotfiles-root.sh
+    pri "Installing dotfiles for root"
+    sh $DOTFILES_DIR/install-dotfiles-root.sh
+fi
 
 pri "Copying doas configuration"
 cp $CONFIGD_DIR/doas.conf /etc/doas.conf
@@ -134,7 +135,6 @@ chown root:root /etc/doas.conf
 chmod -rw /etc/doas.conf
 
 pri "Installing dmenu_run_history"
-wget --quiet https://tools.suckless.org/dmenu/scripts/dmenu_run_with_command_history/dmenu_run_history -O /bin/dmenu_run_history
 
 pri "Cleaning up"
 rm -rf $USER_HOME/.cargo
